store/controllers/product.py
from fastapi import APIRouter, status, Body, Depends, Path
from pydantic import UUID4

from store.schemas.product import ProductIn, ProductOut
from store.usecases.product import ProductUseCase

router = APIRouter(tags=["products"])


@router.post(path="/", status_code=status.HTTP_201_CREATED)
async def post(
    body: ProductIn = Body(...), usecase: ProductUseCase = Depends()
) -> ProductOut:
    return await usecase.create(body=body)


@router.get(path="/{id}", status_code=status.HTTP_200_OK)
async def get(
    id: UUID4 = Path(alias="id"), usecase: ProductUseCase = Depends()
) -> ProductOut:
    return await usecase.get(id=id)
